fix: resolve merge conflict in code explorer
